% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliabilities.R
\name{alpha}
\alias{alpha}
\alias{alpha_std}
\title{Variants of coefficient alpha}
\usage{
alpha(S)

alpha_std(S)
}
\arguments{
\item{S}{A positive definite covariance matrix.}
}
\value{
Cofficient alpha, standardized, orsigma coefficient alpha.
}
\description{
The ordinary, standardized and sigma coefficient alpha.
}
