% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliabilities.R
\name{bias}
\alias{bias}
\title{The bias term for the weighted alpha.}
\usage{
bias(lambda, sigma, w = rep(1, length(lambda)))
}
\arguments{
\item{lambda}{Vector of loadings.}

\item{sigma}{Vector of standard deviations.}

\item{w}{Vector of weights.}
}
\value{
The bias term, i.e. B such that omega = alpha + B.
}
\description{
The bias term is described in the paper.
}
