% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{error_beta}
\alias{error_beta}
\title{Error distributions.}
\usage{
error_beta(n)
}
\arguments{
\item{n}{Number of samples.}
}
\value{
Random variates.
}
\description{
Three error distributions, all with mean 0 and variance 1. Rescaled variants
of the Beta distribution, Gamma distriution, and t distribution.
}
