% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.distribution.R
\name{fit.distribution}
\alias{fit.distribution}
\title{Calculate distribution parameters and statistics}
\usage{
fit.distribution(data, distr, method = c("mle", "lmom"), na.thres = 10)
}
\arguments{
\item{data}{vector of data}

\item{distr}{character string specifying the distribution, see details}

\item{method}{distribution fitting method, see details}

\item{na.thres}{maximum percentage of NA values allowed in data, default = 10\%}
}
\value{
a named vector containing:
\itemize{
\item{fitted distribution parameters, parameters are distribution-specific}
\item{prob.zero}{empirical probability of zeros in data, used in SI calculation with with distributions not including zero such as gamma and Weibull}
\item{n.obs}{the total number of observations in data (including NA values)}
\item{n.na}{the number of NA values in data}
\item{ks.pval}{p-value for a two-sided Kolmogorov-Smirnov test that data comes form the fitted distribution}
\item{ad.pval}{p-value for a two-sided Anderson-Darling test that data comes form the fitted distribution}
The data (without NA values) and distr, method and na.thres settings are added to the result as addditional attributes.
}
}
\description{
This function calculates the distribution parameters and a number of statistics for a numeric vector and returns them as
a named vector. This function is used internally by the standardized.index function or can be used to provide input to it.
}
\details{
Supported distributions are: gamma ('gamma'), 3-parameter gamma ('gamma3'), Weibull ('weibull'), 3-parameter Weibull ('weibull3'),
Generalized Extreme Value ('gev'), and Generalized Logistic ('glogis'). 
Supported distribution fitting methods are: Maximum Likelihood Estimation ('mle', default for 'gamma','weibull','gev' and 'glogis') and L-Moments ('lmom', default for 'gamma3' and 'weibull3').
'mle' is not supported for distributions 'gamma3' and 'weibull3'. For distr = 'glogis' and method = 'lmom', the 'glo' distribution from package 'lmomco' is used, and its parameters are returned.
}
\examples{
data(Ukkel_RR)
# calculate the total rainfall for all months June
monthly.precipitation <- apply.monthly(x=Ukkel_RR,FUN=sum)
data <- c(coredata(monthly.precipitation[format(index(monthly.precipitation),'\%m')=='06']))
# fit gamma distribution to the data
fit <- fit.distribution(data=data,distr='gamma',method='mle')
fitplot(x=fit,main='June precipitation',xlab='precipitation (mm)')
# fit gev distribution to the data
fit <- fit.distribution(data=data,distr='gev',method='mle')
fitplot(x=fit,main='June precipitation',xlab='precipitation (mm)')
# fit glogis distribution to the data
fit <- fit.distribution(data=data,distr='glogis',method='mle')
fitplot(x=fit,main='June precipitation',xlab='precipitation (mm)')

}
\seealso{
\code{\link{standardized.index}}, \code{\link{fitplot}}
}
