% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitplot.R
\name{fitplot}
\alias{fitplot}
\title{Goodness-of-fit Plot}
\usage{
fitplot(x, date = NULL, main = NULL, xlab = 'data', filename = NULL)
}
\arguments{
\item{x}{output object of the fit.distribution function}

\item{date}{reference date for the distribution parameters (optional)}

\item{main}{plot title (optional)}

\item{xlab}{X-axis label for plot (optional)}

\item{filename}{if provided, the plot is written to a .png file with this filename, otherwise console output is used}
}
\value{
A figure displaying goodness-of-fit information.
}
\description{
Function to generate a goodness-of-fit plot for output of the fit.distribution function.
}
\examples{
data(Ukkel_RR)
# calculate the total rainfall for all months June
monthly.precipitation <- apply.monthly(x=Ukkel_RR,FUN=sum)
data <- c(coredata(monthly.precipitation[format(index(monthly.precipitation),'\%m')=='06']))
# fit gamma distribution to the data?pn
fit <- fit.distribution(data=data,distr='gamma',method='mle')
# goodness of fit plot
fitplot(x=fit,main='June precipitation',xlab='precipitation (mm)')

}
\seealso{
\code{\link{standardized.index}}, \code{\link{fitplot}}
}
