% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.aggregated.value.R
\name{get.aggregated.value}
\alias{get.aggregated.value}
\title{Get Aggregated Value}
\usage{
get.aggregated.value(date, data, agg.length, agg.fun = "sum",
  agg.na.thres = 10, agg.interpolation = c("none", "linear", "mean",
  "zeros"), period.warn = TRUE)
}
\arguments{
\item{date}{date for which the aggregated value is determined}

\item{data}{an xts object containing daily data}

\item{agg.length}{length of the aggregation period in days}

\item{agg.fun}{function on x to apply to the aggregation data, default is 'sum'}

\item{agg.na.thres}{threshold for the percentage of na values allowed in the aggregation period data, default = 10}

\item{agg.interpolation}{interpolation type for missing values in individual aggregation period data before applying agg.fun: 'none' (default, NA's are removed), 'linear', 'mean', or 'zeros'}

\item{period.warn}{if TRUE, a warning is when the requested aggregation period contains dates not included in data}
}
\value{
A numeric value giving the aggregated value for the aggregation period
}
\description{
Function to get the aggregated value for a single aggregation period
}
\details{
This function determines the dates in an aggregation period for standardized index calculation, extracts the corresponding data and
applies an aggregation function to the data (default is 'sum', but any function on x can be passed).
}
\examples{
data(Ukkel_RR)
date <- as.Date('2018-07-01')
# get the aggregated value values for the 30 day-period preceding date
get.aggregated.value(date = date, data = Ukkel_RR, agg.length=30)

}
\seealso{
\code{\link{standardized.index}}, \code{\link{get.reference.values}}
}
