% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glo.R
\name{glo}
\alias{glo}
\alias{pglo}
\alias{rglo}
\title{Generalized Logistic Distribution}
\usage{
pglo(q, xi, alpha, kappa)

rglo(n, xi, alpha, kappa)
}
\arguments{
\item{q}{vector of quantiles}

\item{xi}{distribution location parameter}

\item{alpha}{distribution scale parameter}

\item{kappa}{distribution shape parameter}

\item{n}{number of observatins}
}
\value{
pglo gives the distribution function and rglo generates random deviates.
}
\description{
Generalized Logistic Distribution
}
\details{
Generalized Logistic Distribution CDF definition as a wrapper for cdfglo analogous to other pdistr functions.
}
\examples{
data(Ukkel_RR)
# calculate the total rainfall for all months June
monthly.precipitation <- xts::apply.monthly(x=Ukkel_RR,FUN=sum)
data <- c(coredata(monthly.precipitation[format(index(monthly.precipitation),'\\\%m')=='06']))
# fit generalized logistic distribution to the data
fit <- fit.distribution(data=data,distr='glo',method='lmom')
fit

}
\seealso{
\code{\link{cdfglo}}, \code{\link{fit.distribution}}
}
