% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latcon.R
\name{latcon}
\alias{latcon}
\title{Estimate a Latent Congeneric Model}
\usage{
latcon(data, use = "complete.obs", ...)
}
\arguments{
\item{data}{A data frame of observations or a named list with elements
\code{lambda}, \code{sigma}, and \code{cuts}. See the details..}

\item{use}{Passed to \code{stats::cov}; defaults to \code{"complete.obs"}.}

\item{...}{Passed to \code{psych::fa}, where \code{fm = "ml"} by default.}
}
\value{
An object of class \code{latcon}.
}
\description{
This function uses \code{psych::polychoric} and \code{psych::fa} to do a barebones
multivariate ogive model. It is similar to \code{psych::irt.fa}. The \code{...}
arguments are passed to \code{psych::fa}, which is called with \code{fm = "ml"} by
default.
}
\examples{
extraversion = psychTools::bfi[c("E1", "E2", "E3", "E4", "E5")]
extraversion[, "E1"] = 7 - extraversion[, "E1"] # Reverse-coded item.
fit = latcon(extraversion)
}
