% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliabilities.R
\name{omega}
\alias{omega}
\alias{omega_h}
\alias{omega_sigma}
\alias{omega_std}
\title{Variants of the omega coefficient}
\usage{
omega(lambda, sigma)

omega_h(lambda, sigma)

omega_sigma(lambda, sigma)

omega_std(lambda, sigma)
}
\arguments{
\item{lambda}{Vector of loadings.}

\item{sigma}{Vector of standard deviations.}
}
\value{
The congeneric reliability, coefficient H, standardized reliability,
or sigma reliability.
}
\description{
The congeneric reliability, coefficient H, standardized reliability, and
sigma reliability.
}
