% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal.R
\name{ordinal_omega}
\alias{ordinal_omega}
\alias{ordinal_std}
\alias{ordinal_H}
\title{Theoretical ordinal reliability}
\usage{
ordinal_omega(lambda, sigma, cuts, type = c("H", "std"))

ordinal_std(lambda, sigma, cuts)

ordinal_H(lambda, sigma, cuts)
}
\arguments{
\item{lambda}{Vector of loadings.}

\item{sigma}{Vector of standard deviations.}

\item{cuts}{List of cuts, matrix of cuts, or vector of cuts that will be
repeated.}

\item{type}{One of \code{"H"} and \code{"std"}, defaults to \code{"H"}.}
}
\value{
The population ordinal reliabiltiy
}
\description{
Theoretical ordinal reliability
}
