% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal.R
\name{ordinal_poly}
\alias{ordinal_poly}
\title{Empirical ordinal omega}
\usage{
ordinal_poly(
  poly,
  y,
  xi_type = c("theoretical", "sample"),
  type = c("H", "std"),
  ...
)
}
\arguments{
\item{poly}{A an object of class \code{poly}.}

\item{y}{Optional array of observations.}

\item{xi_type}{One of \code{"theoretical"} and \code{"sample"}, defaults to
\code{"theoretical"}. The option \code{"sample"} needs \code{y}}

\item{type}{One of \code{"H"} and \code{"std"}, defaults to \code{"H"}.}

\item{...}{Passed to \code{psych::fa}.}
}
\value{
The empirical ordinal reliability.
}
\description{
Empirical ordinal omega
}
