% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{prettify}
\alias{prettify}
\title{Make pretty strings.}
\usage{
prettify(x, digits = 3, nsmall = 3)
}
\arguments{
\item{x}{Numeric matrix.}

\item{digits}{Digits to prettify with.}

\item{nsmall}{Smallest number of digits to keep.}
}
\value{
Pretty version of the digits.
}
\description{
Make pretty strings.
}
