% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{rescale}
\usage{
rescale(x)
}
\arguments{
\item{x}{a vector of numerric or integer values}
}
\value{
a vector of numerric values
}
\description{
\code{rescale} returns a scaled copy of a vector of values. To do this, it
vector from the value by the standard deviation.
}
\examples{
a <- 50:100
rescale(a)

}
