% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_MSE}
\alias{simulate_MSE}
\title{Simulate MSE for alpha and standardized alpha for a parallel model.}
\usage{
simulate_MSE(nreps = 10^5, n = 100, k = 5, sigma = 1, error = stats::rnorm)
}
\arguments{
\item{nreps}{Number of reptitions.}

\item{n}{Number of observations.}

\item{k}{Size of matrix or number of testlets.}

\item{sigma}{Residual standard deviation.}

\item{error}{Distribution of the error variable.}
}
\value{
Covariance matrix.
}
\description{
The parameter lambda is fixed to one and the latent variable is unit normal.
}
