% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation}
\alias{simulation}
\title{Run the simulation study of the paper.}
\usage{
simulation(nreps)
}
\arguments{
\item{nreps}{Number of repetitons.}
}
\value{
A matrix of alphas and standardized alphas presented in the same way
as in the the paper.
}
\description{
Run the simulation study of the paper.
}
