% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{standardize}
\usage{
standardize(x)
}
\arguments{
\item{x}{a vector of numerric or integer values}
}
\value{
a vector of numerric values
}
\description{
\code{standardize} returns a standardized copy of a vector of values. To do this, it
subtracts the mean of the vector from the value and divide the results by the
the standard deviation.
}
\examples{
a <- 50:100
standardize(a)

}
