% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standaRdized-package.R
\docType{package}
\name{standaRdized-package}
\alias{standaRdized-package}
\alias{standaRdized}
\title{Standardized Index Calculation}
\description{
Functions and methods for the calculation of Standardized Index values
}
\details{
This package provides functions and methods for the calculation of daily Standardized Index values such as the Standardized Precipitation Index (SPI), Standardized Precipitation Evapotranspiration Index (SPEI) or Standardized Streamflow Index (SSI)).
}
\author{
Willem Maetens

Maintainer: Willem Maetens <w.maetens@vmm.be>
}
\keyword{SPEI}
\keyword{SPI}
\keyword{SSI}
\keyword{standardized_index}
\keyword{xts}
