% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardized.index.R
\name{standardized.index}
\alias{standardized.index}
\title{Standardized Index Calculation}
\usage{
standardized.index(data, index.out, agg.length, agg.fun = "sum",
  ref.data = data, distr = c("gamma", "gamma3", "weibull", "weibull3",
  "gev", "glogis"), method = c("mle", "lmom"), params = NULL,
  ks.thres = NULL, ad.thres = NULL, ref.years = NULL,
  ref.length = 30, ref.na.thres = 10, agg.na.thres = ref.na.thres,
  agg.interpolation = c("none", "linear", "mean", "zeros"), digits = 2,
  output.attrs = c("some", "all"))
}
\arguments{
\item{data}{an xts object containing observed daily data}

\item{index.out}{vector of dates for which the standardized index is to be calculated}

\item{agg.length}{length of the aggregation period in days}

\item{agg.fun}{function on x to apply to the aggregation data, default is 'sum'}

\item{ref.data}{an xts object containing daily reference data, default is data itself}

\item{distr}{name of the distribution to be fitted, see details}

\item{method}{distribution fitting method, 'mle' (default) for maximum likelihood estimation, 'lmom' for L-moments}

\item{params}{xts containing the reference distribution parameters for each index.out, if not specified (NULL) or missing for specific dates, they are calculated, in which case ref.data should be specified}

\item{ks.thres}{threshold p-value for the Kolmogrov-Smirnov test, if rejected, the value is set to NA, default= NULL (test not applied)}

\item{ad.thres}{threshold p-value for the Anderson-Darling test, if rejected, the value is set to NA, default= NULL (test not applied)}

\item{ref.years}{years to be used as reference period, default is to use all years in ref.data. If NULL, ref.length is used}

\item{ref.length}{if ref.years is null, the ref.length number of years (default = 30) preceding (but not including) the index.out timestamp are used as reference period}

\item{ref.na.thres}{threshold for the percentage of NA values allowed in reference period data, default = 10\%}

\item{agg.na.thres}{threshold for the percentage of na values allowed in the aggregation period data, default = ref.na.thres}

\item{agg.interpolation}{interpolation type for missing values in aggregation data for an individual reference period element: 'none' (default, NA's are removed), 'linear', 'mean', or 'zeros'}

\item{digits}{number of digits for rounding the resulting standardized index values, default = 2}

\item{output.attrs}{determines what is attached as xtsAttributes: 'some' (default) adds function settings, and 'all' adds function settings, data and ref.data attributes and the reference values for index.out, aggregation values for index.out, and an xts object containing the fitted parameters}
}
\value{
An xts object containing the standardized index values.
}
\description{
Function to calculate a Standardized Index (SPI, SPEI, SSI,...) for a time series.
}
\details{
The argument distr can be either: 'gamma' for the gamma distribution, 'gamma3' for the 3-parameter gamma distribution, 'weibull' for the 'weibull distribution,
'weibull3' for the 'weibull distribution, 'gev' for the Generalized Extreme Value distribution, or 'glo' for the shifted log-logistic distribution.
The \link{fit.distribution} function is used internally to calculate distribution fits, alternatively the parameters for the reference distribution can be supplied via the params argument.
}
\examples{
data(Ukkel_RR)
# since this is rainfall data, we are calculating the SPI
# calculate SPI-1 for July 2011, which is approximated by setting agg.length to 30 days
SPI_1 <- standardized.index(data=Ukkel_RR,agg.length=30,index.out=index(Ukkel_RR['2011-07']))
fprint(SPI_1)
# calculate SPI-3 for July 2011, which is approximated by setting agg.length to 90 days
SPI_3 <- standardized.index(data=Ukkel_RR,agg.length=90,index.out=index(Ukkel_RR['2011-07']))
fprint(SPI_3)

}
\seealso{
\code{\link{fit.distribution}}, \code{\link{get.reference.values}}, \code{\link{fprint}}
}
