% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R, R/reliabilities.R
\name{thurstone}
\alias{thurstone}
\title{Thurstone weights}
\usage{
thurstone(lambda, sigma)

thurstone(lambda, sigma)
}
\arguments{
\item{lambda}{Vector of loadings.}

\item{sigma}{Vector of standard deviations.}
}
\value{
The Thurstone weights.

The Thurstone weights.
}
\description{
Thurstone weights

Thurstone weights
}
