% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{trim_vector}
\alias{trim_vector}
\title{Remove infinities from vector and append and prepend -Inf and Inf.}
\usage{
trim_vector(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Fixed vector.
}
\description{
Remove infinities from vector and append and prepend -Inf and Inf.
}
