% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standaRdized-data.R
\docType{data}
\name{Ukkel_RR}
\alias{Ukkel_RR}
\title{Ukkel Daily Precipitation}
\format{An xts object on 1880-01-01/2019-02-28 containing:
\tabular{llll}{
[,'value'] \tab num \tab Precipitation \tab (mm/day)\cr
}
With xts attributes:
\tabular{lll}{
$ name     : \tab chr \tab "Ukkel" \cr
$ country  : \tab chr \tab "Belgium" \cr
$ element  : \tab chr \tab "RR" \cr
$ unit     : \tab chr \tab "mm" \cr
$ longitude: \tab num \tab 4.36638889 \cr
$ latitude : \tab num \tab 50.8 \cr
$ elevation: \tab num \tab 100 \cr
$ source   : \tab chr \tab "ECA\&D (ecad.eu)" \cr
}}
\source{
Royal Meteorological Institute of Belgium (RMI) via \url{https://www.ecad.eu/}
}
\description{
Daily precipitation (mm/day) for Ukkel from 1898 to 2002.
}
\details{
xts object containing daily rainfall at the Ukkel station in Belgium
}
\examples{
data(Ukkel_RR)
str(Ukkel_RR)
fprint(Ukkel_RR)
}
\keyword{datasets}
