% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{vbeta}
\alias{vbeta}
\title{Variance of the beta distribution}
\usage{
vbeta(shape1, shape2)
}
\arguments{
\item{shape1, shape2}{Shape parameters of the beta density.}
}
\value{
The variance of the beta distribution.
}
\description{
Variance of the beta distribution
}
