% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{vgamma}
\alias{vgamma}
\title{Variance of the gamma distribution}
\usage{
vgamma(shape, ratio)
}
\arguments{
\item{shape, ratio}{Shape parameters of the gamma density.}
}
\value{
The variance of the gamma distribution.
}
\description{
Variance of the gamma distribution
}
