% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{Xi_sample}
\alias{Xi_sample}
\title{Calculate sample XI}
\usage{
Xi_sample(y, cuts)
}
\arguments{
\item{y}{Array of observations.}

\item{cuts}{A matrix, list, or vector of cuts}
}
\value{
Sample Xi matrix.
}
\description{
Calculate sample XI
}
