% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{Xi_theoretical}
\alias{Xi_theoretical}
\title{Calculate the theoretical Xi}
\usage{
Xi_theoretical(cuts, Sigma)
}
\arguments{
\item{cuts}{A matrix, list, or vector of cuts}

\item{Sigma}{The polychoric correlation matrix.}
}
\value{
Theoretical Xi matrix.
}
\description{
Calculate the theoretical Xi
}
