% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latcon.R, R/ordinal.R
\name{ordinal_omega}
\alias{ordinal_omega}
\alias{ordinal_omega_}
\title{Calculate the Ordinal Alpha}
\usage{
ordinal_omega(
  object,
  xi = c("sample", "theoretical"),
  weights = c("optimal", "equal"),
  limit = FALSE
)

ordinal_omega_(lambda, sigma, cuts, type = c("H", "std"))
}
\arguments{
\item{object}{An object of class \code{latcon}.}

\item{xi}{One of \code{"theoretical"} and \code{"sample"}, defaults to \code{"sample"}.}

\item{weights}{One of \code{"optimal"} and \code{"equal"}, defaults to \code{"optimal"}.}

\item{limit}{If \code{TRUE}, calculates the limit ordinal omega.}

\item{lambda}{Vector of loadings.}

\item{sigma}{Vector of standard deviations.}

\item{cuts}{List of cuts, matrix of cuts, or vector of cuts that will be
repeated.}

\item{type}{One of \code{"H"} and \code{"std"}, defaults to \code{"H"}.}
}
\value{
The ordinal omega.

The population ordinal reliabiltiy
}
\description{
Calculate the Ordinal Alpha

Theoretical ordinal reliability
}
